requirements.txt
Django==1.10.2
django-filter==0.15.3
django-widget-tweaks==1.4.1
djangorestframework==3.5.1
Markdown==2.6.7
